Finishing code on owner actions and fixed error button styling

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Product
 from .forms import ProductForm
 
@@ -28,8 +29,12 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a room to studio options """
+    if not request.user.is_superuser:
+        messages.error(request, 'Reserved for In Tune Studio use!')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -49,8 +54,12 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit details of a room """
+    if not request.user.is_superuser:
+        messages.error(request, 'Reserved for In Tune Studio use!')
+        return redirect(reverse('home'))
     
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
@@ -74,8 +83,12 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a room """
+    if not request.user.is_superuser:
+        messages.error(request, 'Reserved for In Tune Studio use!')
+        return redirect(reverse('home'))
     
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
